requirements.txt
Flask
Flask-SQLAlchemy
flask-jwt-extended
python-dotenv
pytest
gunicorn
cryptography
pymysql
